requirements.txt
APScheduler==3.0.0
Flask-MySQLdb==0.2.0
Flask-OAuth==0.12
Flask==0.10.1
Jinja2==2.8
MarkupSafe==0.23
MySQL-python==1.2.5
Werkzeug==0.9.4
boto==2.46.1
gunicorn>=17.5
itsdangerous==0.23
oauth2==1.9.0.post1
oauth2client==1.5.2
oauthlib==1.1.2
pymongo==3.7.0
requests-oauthlib==0.6.1
requests==2.10.0
tweepy==3.6.0
wheel==0.24.0
wsgiref==0.1.2
